Custom exceptions were added
These exceptions are runtime exceptions and they already replaced RuntimeException objects in InsuranceCompanyService

The exceptions are handled by GlobalExceptionHandler that is a
ControllerAdice containing a few methods annotated by ExceptionHandler
annotation

GlobalExceptionHandler contains IncorrectDataContainer that stores
occurred errors. It has public field 'errors' because I guess Jackson
json library needs to get this field to convert the object into json
format. This can be public because IncorrectDataContainer is a private
class in the GlobalExceptionHandler so no one can access this object and
its fields/methods