Raw JPQL support was added
JpqlRestController handles requests with JPQL queries. An operation of
the JPQL query is resolved by a JpqlOperationResolver then the operation
is being executed using an EntityManager instance. This process is
being performed in a JpqlService that then returns a result of the
execution

This support is raw because any changes are not audited by Hibernate
Envis (this tool was added by the way) for some reason. I guess because of direct using an EntityManager
instance instead of a JPA Repository

Other changes: each Rest Controller was moved to the http/controller/rest package