Raw 2.0 JPQL support was improved
Entities that will be modified by jpql queries is audited now. But there
are a few issues that occur because of auditing. For auditing a sql select
query is created according to received jpql to get needed entities and then modify them
one by one in order the Hibernate Envers can notice the changes and insert records about it into the tablse.
Thus a lot of queries occur to modify entities because they are modified
one by one as I said before

Another issue is that the app can handle just one alias for an entity called 'i'. I will try to fix this in the near future.
There is a bad code that need to be refactored as well

I will manage these issue in the near future

Quick guide how the jpql query execution works: the app receive a
JpqlQuery instance, then it is passed into a JpqlService that uses a
JpqlRepositoryResolver to resolve which JpqlRepository the app should
use to execute the query. Once the repository has been found it calls
executeQuery() method in which a JpqlOperation is resolved and then it
invokes executQuery() method that contains logic for performing jpql
queries

The code in the JpqlOperationUtil is pretty bad in my opinion, so it
should be refactored soon