JPQL support was successfully completed*
The main logic of hanfling jqpl queries is contained in a
JpqlOperationUtil object. It has been refactored so it looks nice (at
least looks a bit better) now

JpqlQuery was edited. We dont have to pass 'repositoryName' now, it was
redudant and now we using received jpql query to resolve which
repository should perform the query

A few exception were added as well

* - JQPL support was successfully completed at least for simple queries,
if we need to perform more complex queries for the beggining we have
to test the app